refactor(wizard): extract helper components into wizard/ kin files
Phase 6 second slice. Pull standalone helper components out of
Wizard.tsx into dedicated files. Each is small enough to be
independently testable and avoids cross-import with the step
components that consume them.

Extracted:
  - wizard/components.tsx: ArrowIcon, CheckIcon, EnterHint, ProgressDots
  - wizard/NexSignupPanel.tsx: optional Nex registration panel rendered
    inside IdentityStep (calls broker /nex/register)
  - wizard/LocalLLMPicker.tsx: revealed when "Run a local model" tile
    is on; pulls runtime status from getLocalProvidersStatus
  - wizard/ApiKeyRow.tsx: per-provider CLI-vs-API-key toggle
    (re-exported from Wizard.tsx for back-compat with existing
    onboarding/ApiKeyRow.test.tsx import path)

LOCAL_PROVIDER_LABELS moved to wizard/constants.ts since the SetupStep
tile and the ReadyStep summary both need it; previously the constant
was private to LocalLLMPicker and would have forced 3-way drift.

  - web/src/components/onboarding/Wizard.tsx: -431 LOC (2412 -> 1981)
  - web/src/components/onboarding/wizard/ApiKeyRow.tsx: +81 LOC (new)
  - web/src/components/onboarding/wizard/components.tsx: +77 LOC (new)
  - web/src/components/onboarding/wizard/LocalLLMPicker.tsx: +191 LOC (new)
  - web/src/components/onboarding/wizard/NexSignupPanel.tsx: +106 LOC (new)
  - web/src/components/onboarding/wizard/constants.ts: +14 LOC (LOCAL_PROVIDER_LABELS)

No behavior change. Typecheck clean. All 565 web tests pass. Build clean.

Cumulative Phase 6:
  2641 -> 2412 (types/constants)   -229
       -> 1981 (helper components) -431
  Net: -660 LOC; Wizard.tsx shrunk by 25%.

Next slice: extract the 7 step components (Welcome, Templates, Identity,
Team, Setup, Task, Ready) into wizard/Step*.tsx files.
